+++ packages/account-api-sdk/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sophon-labs/account-api-sdk
 
+## 2.1.6-next.1
+
+### Patch Changes
+
+- Adding social network to redirect url
+- Updated dependencies
+  - @sophon-labs/account-core@2.1.6-next.1
+
 ## 2.1.6-next.0
 
 ### Patch Changes