+++ packages/account-react-native/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @sophon-labs/account-react-native
 
+## 2.1.6-next.1
+
+### Patch Changes
+
+- Adding social network to redirect url
+- Updated dependencies
+  - @sophon-labs/account-message-bridge@2.1.6-next.1
+  - @sophon-labs/account-communicator@2.1.6-next.1
+  - @sophon-labs/account-provider@2.1.6-next.1
+  - @sophon-labs/account-core@2.1.6-next.1
+
 ## 2.1.6-next.0
 
 ### Patch Changes